restructure zoo
